LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/pseudo-elements-013-expected.txt


PASS Initial text-decoration-thickness for highlight pseudo
FAIL text-decoration-thickness for highlight pseudo depending on container assert_equals: expected "30px" but got "0px"
FAIL text-decoration-thickness for highlight pseudo depending on container only defined in a query assert_equals: expected "30px" but got "auto"
